+++ EpsilonScript/AST/SequenceNode.cs
@@ -8,21 +8,34 @@ namespace EpsilonScript.AST
   {
     private Node _leftNode;
     private Node _rightNode;
+    private bool _isSingleNode;
 
-    public override bool IsConstant => _leftNode.IsConstant && _rightNode.IsConstant;
+    public override bool IsConstant =>
+      _isSingleNode ? _rightNode.IsConstant : (_leftNode.IsConstant && _rightNode.IsConstant);
 
     public override void Build(Stack<Node> rpnStack, Element element, Compiler.Options options,
       IVariableContainer variables, IDictionary<VariableId, CustomFunctionOverload> functions)
     {
-      if (!rpnStack.TryPop(out _rightNode) || !rpnStack.TryPop(out _leftNode))
+      if (!rpnStack.TryPop(out _rightNode))
       {
         throw new ParserException(element.Token, "Cannot find tokens to sequence");
+      }
+
+      // Handle trailing semicolon - if there's only one operand, treat it as a no-op
+      if (!rpnStack.TryPop(out _leftNode))
+      {
+        _isSingleNode = true;
+        _leftNode = null;
       }
     }
 
     public override void Execute(IVariableContainer variablesOverride)
     {
-      _leftNode.Execute(variablesOverride);
+      if (!_isSingleNode)
+      {
+        _leftNode.Execute(variablesOverride);
+      }
+
       _rightNode.Execute(variablesOverride);
       ValueType = _rightNode.ValueType;
 
@@ -38,6 +51,11 @@ namespace EpsilonScript.AST
       {
         Execute(null);
         return CreateValueNode();
+      }
+
+      if (_isSingleNode)
+      {
+        return _rightNode.Optimize();
       }
 
       _leftNode = _leftNode.Optimize();